create modulo xml and migrate name from dynamicView to CraftD
