+++ app/build.gradle
@@ -50,6 +50,7 @@ dependencies {
     def koin_version = "2.2.3"
     implementation(project(":dynamic"))
     implementation(project(":craftd-core"))
+    implementation(project(":craftd-xml"))
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
