+++ app/src/main/java/com/github/rviannaoliveira/dynamicview/xml/presentation/DynamicComponentActivity.kt
@@ -3,7 +3,7 @@ package com.github.rviannaoliveira.dynamicview.xml.presentation
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.github.rviannaoliveira.dynamic.xml.presentation.ui.CraftDViewAdapter
+import com.github.codandotv.craftd.xml.ui.CraftDViewAdapter
 import com.github.rviannaoliveira.dynamicview.app.databinding.ActivityDynamicComponentBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -16,7 +16,7 @@ class CraftDComponentActivity : AppCompatActivity() {
         binding = ActivityDynamicComponentBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        binding.recycler.adapter = vm.dynamic as CraftDViewAdapter
+        binding.recycler.adapter = vm.dynamic as com.github.codandotv.craftd.xml.ui.CraftDViewAdapter
 
         vm.deeplink.observe(this) {
             Toast.makeText(this, it, Toast.LENGTH_LONG).show()
