+++ craftd-xml/src/main/java/com/github/codandotv/craftd/xml/ui/CraftDEmptyViewHolder.kt
@@ -1,4 +1,4 @@
-package com.github.rviannaoliveira.dynamic.xml.presentation.ui
+package com.github.codandotv.craftd.xml.ui
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
