+++ dynamic/src/main/java/com/github/rviannaoliveira/dynamic/compose/presentation/ui/text/CraftDText.kt
@@ -12,9 +12,9 @@ import androidx.compose.ui.unit.sp
 import androidx.core.text.HtmlCompat
 import com.github.codandotv.craftd.androidcore.data.model.text.TextProperties
 import com.github.codandotv.craftd.androidcore.extensions.empty
+import com.github.rviannaoliveira.dynamic.compose.presentation.extensions.parseColorCompose
 import com.github.rviannaoliveira.dynamic.compose.presentation.ui.toAlignCompose
 import com.github.rviannaoliveira.dynamic.compose.presentation.ui.toTextStyle
-import com.github.rviannaoliveira.dynamic.xml.presentation.ui.button.parseColorCompose
 
 @Composable
 fun DynamicText(
